feat(web-deployment.yaml, values.yaml): add support for configurable endpoint for liveness and readiness probes
The liveness and readiness probes for the web deployment can now be configured with a custom endpoint by setting the `endpoint` property in the `httpGet` object of the `livenessProbe` and `readinessProbe` objects in the `values.yaml` file. This allows for more flexibility in configuring the probes to work with different applications.